requirements.txt
PyQt6==6.2.2
paho-mqtt==1.6.1
